--- src/components/organisms/app/AppHeader/AppHeader.stories.ts
@@ -1,15 +1,0 @@
-import AppHeaderComponent from './AppHeader';
-
-import { Meta, StoryObj } from '@storybook/react';
-
-const meta: Meta<typeof AppHeaderComponent> = {
-  component: AppHeaderComponent,
-};
-
-export default meta;
-
-type Story = StoryObj<typeof AppHeaderComponent>;
-
-export const Basic: Story = {
-  args: {},
-};
